Move ArchiveId class to metadata module
And Improve imports
