+++ aqt/installer.py
@@ -32,11 +32,8 @@ import time
 from logging import getLogger
 from logging.handlers import QueueHandler
 
-from semantic_version import Version
-
 import aqt
 from aqt.archives import QtArchives, SrcDocExamplesArchives, ToolArchives
-from aqt.metadata import ListCommand
 from aqt.exceptions import (
     ArchiveConnectionError,
     ArchiveDownloadError,
@@ -44,13 +41,13 @@ from aqt.exceptions import (
     NoPackageFound,
 )
 from aqt.helper import (
-    ArchiveId,
     MyQueueListener,
     Settings,
     downloadBinaryFile,
     getUrl,
     setup_logging,
 )
+from aqt.metadata import ArchiveId, ListCommand, Version
 from aqt.updater import Updater
 
 try:
