+++ aqt/metadata.py
@@ -26,25 +26,29 @@ import random
 import re
 from logging import getLogger
 from typing import (
-    List,
-    Optional,
-    Union,
-    Tuple,
-    Iterable,
+    Callable,
     Dict,
     Generator,
+    Iterable,
     Iterator,
-    Callable,
+    List,
+    Optional,
+    Tuple,
+    Union,
 )
 from xml.etree import ElementTree as ElementTree
 
 import bs4
+from semantic_version import SimpleSpec as SementicSimpleSpec
 from semantic_version import Version as SemanticVersion
-from semantic_version import SimpleSpec
 from texttable import Texttable
 
-from aqt.exceptions import CliInputError, ArchiveConnectionError, ArchiveDownloadError
-from aqt.helper import ArchiveId, xml_to_modules, Settings, getUrl
+from aqt.exceptions import ArchiveConnectionError, ArchiveDownloadError, CliInputError
+from aqt.helper import Settings, getUrl, xml_to_modules
+
+
+class SimpleSpec(SementicSimpleSpec):
+    pass
 
 
 class Version(SemanticVersion):
@@ -186,6 +190,80 @@ def get_semantic_version(qt_ver: str, is_preview: bool) -> Optional[Version]:
     elif len(qt_ver) == 2:
         return Version(major=int(qt_ver[:1]), minor=int(qt_ver[1:2]), patch=0)
     raise ValueError("Invalid version string '{}'".format(qt_ver))
+
+
+class ArchiveId:
+    CATEGORIES = ("tools", "qt5", "qt6")
+    HOSTS = ("windows", "mac", "linux")
+    TARGETS_FOR_HOST = {
+        "windows": ["android", "desktop", "winrt"],
+        "mac": ["android", "desktop", "ios"],
+        "linux": ["android", "desktop"],
+    }
+    ALL_EXTENSIONS = (
+        "wasm",
+        "src_doc_examples",
+        "preview",
+        "wasm_preview",
+        "x86_64",
+        "x86",
+        "armv7",
+        "arm64_v8a",
+    )
+
+    def __init__(self, category: str, host: str, target: str, extension: str = ""):
+        if category not in ArchiveId.CATEGORIES:
+            raise ValueError("Category '{}' is invalid".format(category))
+        if host not in ArchiveId.HOSTS:
+            raise ValueError("Host '{}' is invalid".format(host))
+        if target not in ArchiveId.TARGETS_FOR_HOST[host]:
+            raise ValueError("Target '{}' is invalid".format(target))
+        if extension and extension not in ArchiveId.ALL_EXTENSIONS:
+            raise ValueError("Extension '{}' is invalid".format(extension))
+        self.category: str = category
+        self.host: str = host
+        self.target: str = target
+        self.extension: str = extension
+
+    def is_preview(self) -> bool:
+        return "preview" in self.extension if self.extension else False
+
+    def is_qt(self) -> bool:
+        return self.category.startswith("qt")
+
+    def is_tools(self) -> bool:
+        return self.category == "tools"
+
+    def is_no_arch(self) -> bool:
+        """Returns True if there should be no arch attached to the module names"""
+        return self.extension in ("src_doc_examples",)
+
+    def is_major_ver_mismatch(self, qt_version: Version) -> bool:
+        """Returns True if the version specifies a version different from the specified category"""
+        return self.is_qt() and int(self.category[-1]) != qt_version.major
+
+    def to_url(self, qt_version_no_dots: Optional[str] = None, file: str = "") -> str:
+        base = "online/qtsdkrepository/{os}{arch}/{target}/".format(
+            os=self.host,
+            arch="_x86" if self.host == "windows" else "_x64",
+            target=self.target,
+        )
+        if not qt_version_no_dots:
+            return base
+        folder = "{category}_{ver}{ext}/".format(
+            category=self.category,
+            ver=qt_version_no_dots,
+            ext="_" + self.extension if self.extension else "",
+        )
+        return base + folder + file
+
+    def __str__(self) -> str:
+        return "{cat}/{host}/{target}{ext}".format(
+            cat=self.category,
+            host=self.host,
+            target=self.target,
+            ext="" if not self.extension else "/" + self.extension,
+        )
 
 
 class ListCommand:
